fixed 'json: cannot unmarshal number .... into Go struct field OpenOrder.orderNumber of type int' for 32 bit systems

+++ order.go
@@ -21,7 +21,7 @@ type Orderb struct {
 */
 
 type OpenOrder struct {
-	OrderNumber int     `json:"orderNumber,string"`
+	OrderNumber int64   `json:"orderNumber,string"`
 	Type        string  `json:"type"`
 	Rate        float64 `json:"rate,string"`
 	Amount      float64 `json:"amount,string"`
